+++ ZoomSDK/ZoomSDK.framework/Versions/A/Headers/ZoomSDKMeetingUIController.h
@@ -8,8 +8,9 @@
 @protocol ZoomSDKMeetingUIControllerDelegate <NSObject>
 /**
  * @brief Designated for Zoom Meeting notify the Invite button be clicked by user if u set showInviteButtonInToolbar =YES;
+ * @param show, default show origin zoom invite window, if u don't want original action, u can set *show = N0, and do UI action yourself in this callback
  */
-- (void)onToolbarInviteButtonClick;
+- (void)onToolbarInviteButtonClick:(BOOL*)show;
 
 /**
  * @brief Designated for Zoom Meeting notify participants click leave meeting.