+++ ZoomSDKSample/ZoomSDKSample/AppDelegate.m
@@ -66,7 +66,7 @@
     _hasLogined = NO;
     _selectDeviceType = H323DeviceType_H323;
     _screenType = ScreenType_First;
-    [[ZoomSDK sharedSDK]initSDK:YES];
+    [[ZoomSDK sharedSDK]initSDK:NO];
     ZoomSDKNetworkService* networkService = [[ZoomSDK sharedSDK] getNetworkService];
     networkService.delegate = self;
     //video window init
@@ -1718,9 +1718,10 @@ void processSignal(int num)
     _shareStatusMsgView.string = info;
 }
 
-- (void)onToolbarInviteButtonClick
+- (void)onToolbarInviteButtonClick:(BOOL*)show
 {
     NSLog(@"Invite Button Clicked by User, Do something!!!");
+    //*show = NO;
 }